// main/src/comment/comment.interface.ts
/**
 * This file is auto-generated by nestjs-proto-gen-ts
 */

import { Metadata } from '@grpc/grpc-js';

export default interface CommentInterface {
  getAllCommentById(
    data: GetAllCommentByIdRequest,
    metadata?: Metadata,
    ...rest: any[]
  ): Promise<GetAllCommentByIdResponse>;
  addCommentToPost(
    data: AddCommentToPostRequest,
    metadata?: Metadata,
    ...rest: any[]
  ): Promise<AddCommentToPostResponse>;
}
// tslint:disable-next-line:no-empty-interface
// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface GetAllCommentByIdRequest {
  postId: string;
}
export interface GetAllCommentByIdResponse {
  data: {
    postId: string;
    authorId: Object;
    content: string;
  }[];
}

export interface AddCommentToPostRequest {
  postId: string;
  authorId: string;
  content: string;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface AddCommentToPostResponse {
  data: {
    postId: string;
    authorId: string;
    content: string;
  };
}